Unify keyboard_default on all platforms
