+++ src/macosx/MacMiniFB.m
@@ -3,6 +3,7 @@
 #include "WindowData_OSX.h"
 #include <MiniFB.h>
 #include <MiniFB_enums.h>
+#include <MiniFB_internal.h>
 #include <Cocoa/Cocoa.h>
 #if defined(USE_METAL_API)
 #include <Carbon/Carbon.h>
